Create container ploigos-jenkins-plugins that bundles Ploigos Jenkins plugins.
This container is built with a two-stage process. The first uses the existing Jenkins image so as to use the
install-plugins.sh script to pull down the plugins listed in ploigos-plugins.txt. The second stage copies all
of these plugins into a ubi-minimal container.

The container can then be used as an initContainer for a Jenkins deployment, with the Jenkins PV mounted into
the container. The content can then be copied into $JENKINS_HOME/plugins.

This container can then be shipped into disconnected environments to provide all the required plugins to
support Ploigos deployment.